fix(ui): Score can be saved with with 0.5 value [EV-14945] (#10252) (#10788)

+++ src/client/author/questionUtils.js
@@ -522,7 +522,7 @@ export const isIncompleteQuestion = (item, itemLevelScoring = false) => {
       if (score === undefined) {
         return [true, "Score needs to be set"];
       }
-      if (parseInt(score, 10) === 0) {
+      if (parseFloat(score, 10) === 0) {
         return [true, "Score cannot be zero"];
       }
     }
